Move label utils to separate files
